requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
astroid==3.3.10
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coverage==7.9.1
croniter==6.0.0
dill==0.4.0
distro==1.9.0
Flask==3.1.0
flask-cors==6.0.1
flask_jwt_extended>=4.4.4
flexcache==0.3
flexparser==0.4
frozenlist==1.7.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.6.0
openai==0.28.0
packaging==25.0
Pint==0.24.4
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pylint==3.3.7
pytest==8.3.5
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
requests==2.31.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stravalib==2.3
tomlkit==0.13.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.1
